fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21556 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBEdge.h
@@ -1127,13 +1127,13 @@ public:
 
     // returns a reference to the internal structure for the convenience of NETEDIT
     Lane& getLaneStruct(int lane) {
-        assert(lane < myLanes.size());
+        assert(lane < (int)myLanes.size());
         return myLanes[lane];
     }
 
     // returns a reference to the internal structure for the convenience of NETEDIT
     const Lane& getLaneStruct(int lane) const {
-        assert(lane < myLanes.size());
+        assert(lane < (int)myLanes.size());
         return myLanes[lane];
     }
